// apps/web/src/lib/api/generated/models/aiAppraisalAnswerValueDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DeepRef API
 * OpenAPI spec version: 0.1.0
 */

export type AiAppraisalAnswerValueDto =
	| {
			kind: 'enum';
			value: string;
	  }
	| {
			kind: 'boolean';
			value: boolean;
	  }
	| {
			kind: 'scale';
			value: number;
	  }
	| {
			kind: 'text';
			value: string;
	  };